changes format datatype

+++ functions_nested_loops/102-fibonacci.c
@@ -12,17 +12,17 @@
  */
 int main(void)
 {
-	int a = 1; 
-	int b = 2; 
-	int next;  
-	int i;
+	long a = 1; 
+	long b = 2; 
+	long next;  
+	long i;
 
-	printf("%d, %d, ", a, b);
+	printf("%lu, %lu, ", a, b);
 
 	for (i = 3; i <= 50; i++)
 	{
 		next = a + b;
-		printf("%d", next);
+		printf("%lu", next);
 		if (i < 50)
 			printf(", ");
 		a = b;
